1. 고양이 에너지 확인 로직 구현 완료(30 이하인지, 0 인지) 2. application/service/EnergyService 추가 3. DB getter 추가 4. CatEntity 및 CatDTO 몸무게 double로 수정 5. CatEntity 및 CatDTO 레벨 추가

+++ src/test/java/com/catholic/meowlife/dto/CatDTO.java
@@ -1,6 +1,8 @@
 package com.catholic.meowlife.dto;
 
+import com.catholic.meowlife.domain.entity.CatEntity;
 import lombok.*;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @AllArgsConstructor
@@ -13,7 +15,7 @@ public class CatDTO {
     private String catName;
     private String catBreeds;
     private int energy;
-    private double weight;
+    private int weight;
     private int exp;
     private int level;
 }
